requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-filter==23.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
openapi-codec==1.3.2
packaging==23.1
prompt-toolkit==3.0.38
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
SQLAlchemy==1.4.47
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
psycopg2==2.9.9
psycopg2-binary==2.9.9